+++ packages/wayfinder-react/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @ar.io/wayfinder-react
 
+## 1.0.15-alpha.2
+
+### Patch Changes
+
+- Updated dependencies [eb839e4]
+  - @ar.io/wayfinder-core@1.4.0-alpha.2
+
+## 1.0.15-alpha.1
+
+### Patch Changes
+
+- Updated dependencies [98d47cd]
+  - @ar.io/wayfinder-core@1.4.0-alpha.1
+
+## 1.0.15-alpha.0
+
+### Patch Changes
+
+- Updated dependencies [c3fc591]
+- Updated dependencies [9fad87b]
+  - @ar.io/wayfinder-core@1.4.0-alpha.0
+
 ## 1.0.14
 
 ### Patch Changes